word bank and beginning of functionality
Added the word bank of the top 50 most popular programming languages from the TIOBE index and some of the more popular JS and python libraries, as well as Databases and tools and such. Also added the first part of functionality, so it can be run and will report back the amount of hits for each word in the wordbank.